+++ package.json
@@ -25,6 +25,7 @@
 		"@google-cloud/storage": "^7.15.2",
 		"@guardian/content-api-models": "^17.8.0",
 		"aws-lambda": "^1.0.7",
+		"axios": "^1.0.0",
 		"body-parser": "^1.20.2",
 		"date-fns": "^2.30.0",
 		"express": "^4.21.2",
