--- printable-recipe-generator/.prettierignore
@@ -1,5 +1,0 @@
-# Add files here to ignore them from prettier formatting
-/dist
-/coverage
-/.nx/cache
-/.nx/workspace-data
