--- printable-recipe-generator/.vscode/extensions.json
@@ -1,8 +1,0 @@
-{
-	"recommendations": [
-		"nrwl.angular-console",
-		"esbenp.prettier-vscode",
-		"dbaeumer.vscode-eslint",
-		"firsttris.vscode-jest-runner"
-	]
-}
