--- printable-recipe-generator/apps/printable-recipe-generator-e2e/package.json
@@ -1,26 +1,0 @@
-{
-	"name": "@printable-recipe-generator/printable-recipe-generator-e2e",
-	"version": "0.0.1",
-	"private": true,
-	"nx": {
-		"name": "printable-recipe-generator-e2e",
-		"implicitDependencies": [
-			"printable-recipe-generator"
-		],
-		"targets": {
-			"e2e": {
-				"executor": "@nx/jest:jest",
-				"outputs": [
-					"{projectRoot}/test-output/jest/coverage"
-				],
-				"options": {
-					"jestConfig": "apps/printable-recipe-generator-e2e/jest.config.ts",
-					"passWithNoTests": true
-				},
-				"dependsOn": [
-					"printable-recipe-generator:build"
-				]
-			}
-		}
-	}
-}
