--- printable-recipe-generator/apps/printable-recipe-generator-e2e/src/printable-recipe-generator/printable-recipe-generator.spec.ts
@@ -1,10 +1,0 @@
-import axios from 'axios';
-
-describe('GET /', () => {
-	it('should return a message', async () => {
-		const res = await axios.get(`/`);
-
-		expect(res.status).toBe(200);
-		expect(res.data).toEqual({ message: 'Hello API' });
-	});
-});
