--- printable-recipe-generator/apps/printable-recipe-generator-e2e/src/support/global-setup.ts
@@ -1,10 +1,0 @@
-/* eslint-disable */
-var __TEARDOWN_MESSAGE__: string;
-
-module.exports = async function () {
-	// Start services that that the app needs to run (e.g. database, docker-compose, etc.).
-	console.log('\nSetting up...\n');
-
-	// Hint: Use `globalThis` to pass variables to global teardown.
-	globalThis.__TEARDOWN_MESSAGE__ = '\nTearing down...\n';
-};
