--- printable-recipe-generator/apps/printable-recipe-generator-e2e/src/support/global-teardown.ts
@@ -1,7 +1,0 @@
-/* eslint-disable */
-
-module.exports = async function () {
-	// Put clean up logic here (e.g. stopping services, docker-compose, etc.).
-	// Hint: `globalThis` is shared between setup and teardown.
-	console.log(globalThis.__TEARDOWN_MESSAGE__);
-};
