--- printable-recipe-generator/apps/printable-recipe-generator-e2e/src/support/test-setup.ts
@@ -1,9 +1,0 @@
-/* eslint-disable */
-import axios from 'axios';
-
-module.exports = async function () {
-	// Configure axios for tests to use.
-	const host = process.env.HOST ?? 'localhost';
-	const port = process.env.PORT ?? '3000';
-	axios.defaults.baseURL = `http://${host}:${port}`;
-};
