--- printable-recipe-generator/apps/printable-recipe-generator-e2e/tsconfig.json
@@ -1,11 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"compilerOptions": {
-		"outDir": "out-tsc/printable-recipe-generator-e2e",
-		"esModuleInterop": true,
-		"noUnusedLocals": false,
-		"noImplicitAny": false
-	},
-	"include": ["jest.config.ts", "src/**/*.ts"],
-	"references": []
-}
