--- printable-recipe-generator/apps/printable-recipe-generator/Dockerfile
@@ -1,23 +1,0 @@
-# This file is generated by Nx.
-#
-# Build the docker image with `npx nx docker-build printable-recipe-generator`.
-# Tip: Modify "docker-build" options in project.json to change docker build args.
-#
-# Run the container with `docker run -p 3000:3000 -t printable-recipe-generator`.
-FROM docker.io/node:22-slim
-
-WORKDIR /app
-
-#RUN addgroup --system printable-recipe-generator && \
-#          adduser --system -G printable-recipe-generator printable-recipe-generator
-
-COPY package.json printable-recipe-generator/
-RUN npm --prefix printable-recipe-generator --omit=dev -f install
-COPY dist printable-recipe-generator/
-#RUN chown -R printable-recipe-generator:printable-recipe-generator .
-
-# You can remove this install step if you build with `--bundle` option.
-# The bundled output will include external dependencies.
-
-
-CMD [ "node", "printable-recipe-generator/main.js" ]
