--- printable-recipe-generator/apps/printable-recipe-generator/package.json
@@ -1,78 +1,0 @@
-{
-	"name": "@printable-recipe-generator/printable-recipe-generator",
-	"version": "0.0.1",
-	"private": true,
-	"nx": {
-		"name": "printable-recipe-generator",
-		"targets": {
-			"build": {
-				"executor": "@nx/esbuild:esbuild",
-				"outputs": [
-					"{options.outputPath}"
-				],
-				"defaultConfiguration": "production",
-				"options": {
-					"platform": "node",
-					"outputPath": "apps/printable-recipe-generator/dist",
-					"format": [
-						"cjs"
-					],
-					"bundle": false,
-					"main": "apps/printable-recipe-generator/src/main.ts",
-					"tsConfig": "apps/printable-recipe-generator/tsconfig.app.json",
-					"assets": [
-						"apps/printable-recipe-generator/src/assets"
-					],
-					"esbuildOptions": {
-						"sourcemap": true,
-						"outExtension": {
-							".js": ".js"
-						}
-					}
-				},
-				"configurations": {
-					"development": {},
-					"production": {
-						"generateLockfile": true,
-						"esbuildOptions": {
-							"sourcemap": false,
-							"outExtension": {
-								".js": ".js"
-							}
-						}
-					}
-				}
-			},
-			"serve": {
-				"executor": "@nx/js:node",
-				"defaultConfiguration": "development",
-				"dependsOn": [
-					"build"
-				],
-				"options": {
-					"buildTarget": "printable-recipe-generator:build",
-					"runBuildTargetDependencies": false
-				},
-				"configurations": {
-					"development": {
-						"buildTarget": "printable-recipe-generator:build:development"
-					},
-					"production": {
-						"buildTarget": "printable-recipe-generator:build:production"
-					}
-				}
-			},
-			"test": {
-				"options": {
-					"passWithNoTests": true
-				}
-			},
-			"docker-build": {
-				"dependsOn": [
-					"build"
-				],
-				"command": "docker build -f apps/printable-recipe-generator/Dockerfile . -t printable-recipe-generator"
-			}
-		}
-	}
-}
