--- printable-recipe-generator/apps/printable-recipe-generator/tsconfig.json
@@ -1,13 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"files": [],
-	"include": [],
-	"references": [
-		{
-			"path": "./tsconfig.app.json"
-		},
-		{
-			"path": "./tsconfig.spec.json"
-		}
-	]
-}
