+++ printable-recipe-generator/jest.config.ts
@@ -1,6 +1,11 @@
-import type { Config } from 'jest';
-import { getJestProjectsAsync } from '@nx/jest';
-
-export default async (): Promise<Config> => ({
-	projects: await getJestProjectsAsync(),
-});
+/* eslint-disable */
+export default {
+	displayName: 'printable-recipe-generator',
+	preset: '../jest.preset.js',
+	testEnvironment: 'node',
+	transform: {
+		'^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
+	},
+	moduleFileExtensions: ['ts', 'js', 'html'],
+	coverageDirectory: '../coverage/printable-recipe-generator',
+};
