--- printable-recipe-generator/jest.preset.js
@@ -1,3 +1,0 @@
-const nxPreset = require('@nx/jest/preset').default;
-
-module.exports = { ...nxPreset };
