--- printable-recipe-generator/package.json
@@ -1,45 +1,0 @@
-{
-	"name": "@printable-recipe-generator/source",
-	"version": "0.0.0",
-	"license": "MIT",
-	"scripts": {},
-	"private": true,
-	"dependencies": {
-		"axios": "^1.6.0",
-		"ejs": "^3.1.10",
-		"express": "^4.21.2",
-		"express-rate-limit": "^7.5.0"
-	},
-	"devDependencies": {
-		"@nx/esbuild": "20.8.0",
-		"@nx/eslint": "20.8.0",
-		"@nx/eslint-plugin": "20.8.0",
-		"@nx/jest": "20.8.0",
-		"@nx/js": "20.8.0",
-		"@nx/node": "20.8.0",
-		"@nx/workspace": "20.8.0",
-		"@swc-node/register": "~1.9.1",
-		"@swc/core": "~1.5.7",
-		"@swc/helpers": "~0.5.11",
-		"@swc/jest": "~0.2.36",
-		"@types/express": "^4.17.21",
-		"@types/jest": "^29.5.12",
-		"@types/node": "~18.16.9",
-		"@typescript-eslint/eslint-plugin": "^7.16.0",
-		"@typescript-eslint/parser": "^7.16.0",
-		"esbuild": "^0.19.2",
-		"eslint": "~8.57.0",
-		"eslint-config-prettier": "^10.0.0",
-		"jest": "^29.7.0",
-		"jest-environment-node": "^29.7.0",
-		"nx": "20.8.0",
-		"prettier": "^2.6.2",
-		"ts-jest": "^29.1.0",
-		"ts-node": "10.9.1",
-		"tslib": "^2.3.0",
-		"typescript": "~5.7.2"
-	},
-	"workspaces": [
-		"apps/*"
-	]
-}
