+++ printable-recipe-generator/src/assets/recipe.ejs
