+++ printable-recipe-generator/tsconfig.app.json
@@ -1,0 +1,10 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"outDir": "../dist/out-tsc",
+		"module": "commonjs",
+		"types": ["node"]
+	},
+	"exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"],
+	"include": ["src/**/*.ts"]
+}
