+++ printable-recipe-generator/tsconfig.json
@@ -1,13 +1,16 @@
 {
-	"extends": "./tsconfig.base.json",
-	"compileOnSave": false,
+	"extends": "../tsconfig.base.json",
 	"files": [],
+	"include": [],
 	"references": [
 		{
-			"path": "./apps/printable-recipe-generator-e2e"
+			"path": "./tsconfig.app.json"
 		},
 		{
-			"path": "./apps/printable-recipe-generator"
+			"path": "./tsconfig.spec.json"
 		}
-	]
+	],
+	"compilerOptions": {
+		"esModuleInterop": true
+	}
 }
