+++ printable-recipe-generator/tsconfig.spec.json
@@ -1,7 +1,8 @@
 {
-	"extends": "../../tsconfig.base.json",
+	"extends": "./tsconfig.json",
 	"compilerOptions": {
-		"outDir": "./out-tsc/jest",
+		"outDir": "../dist/out-tsc",
+		"module": "commonjs",
 		"types": ["jest", "node"]
 	},
 	"include": [
@@ -9,10 +10,5 @@
 		"src/**/*.test.ts",
 		"src/**/*.spec.ts",
 		"src/**/*.d.ts"
-	],
-	"references": [
-		{
-			"path": "./tsconfig.app.json"
-		}
 	]
 }
